+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - When calling `update`, a new dialog allows to remove unresolvable packages if any (e.g. when packages have been renamed or deleted from a channel) ([#24][gui24]).
 - API upgraded to 2.4:
   - added a new `/prompt/confirmation/update/remove-unresolvable-packages` message to `/update`
+  - format changes in `/variants.list` (#7)
 
 [gui24]: https://github.com/memo33/sc4pac-gui/issues/24